refactor(scroll-decks): extract shared 1280x720 canvas frame into pitch/scrollCanvas.jsx
SonarCloud flagged 17.5% new-code duplication on PR #101 because
PitchInvestor.jsx pasted the viewport scaling, HubSpot-widget removal,
and outer canvas chrome straight from PitchShort2.jsx.

This extracts the shared bits into src/pages/pitch/scrollCanvas.jsx:

  - SLIDE_W / SLIDE_H / VIEWPORT_MARGIN constants
  - getScrollDeckScale + useScrollDeckScale (resize + orientationchange)
  - useRemoveHubSpotChatWidget (mutation-observer-backed widget killer)
  - SCROLL_DECK_GLOBAL_CSS (hide HubSpot chrome + cookie banner)
  - ScrollCanvasFrame component (the 1280x720 box, border, brand mark,
    slide counter — child is the actual slide content)

PitchInvestor.jsx now imports these instead of redefining them. The deck
content (the 11 source-deck PNGs slotted into the frame) is unchanged.
